requirements.txt
paymanai
python-dotenv
openai
requests
flask
werkzeug
